feat: auto clean unused files
add documentation info
add metadata to images generated by the plugin
update fixtures screenshots (add metadata)

closes #118
BREAKING CHANGE: To use autocleanup feature you need to update all of the screenshots, best do it by running your test suite with cypress env 'pluginVisualRegressionUpdateImages' set to true.

+++ src/task.hook.ts
@@ -1,11 +1,17 @@
 import fs from "fs";
-import path from "path";
 import { PNG } from "pngjs";
 import pixelmatch, { PixelmatchOptions } from "pixelmatch";
 import moveFile from "move-file";
-import sanitize from "sanitize-filename";
-import { FILE_SUFFIX, IMAGE_SNAPSHOT_PREFIX, TASK } from "./constants";
-import { alignImagesToSameSize, importAndScaleImage } from "./image.utils";
+import { FILE_SUFFIX, TASK } from "./constants";
+import {
+  cleanupUnused,
+  alignImagesToSameSize,
+  scaleImageAndWrite,
+  isImageCurrentVersion,
+  writePNG,
+} from "./image.utils";
+import { generateScreenshotPath } from "./screenshotPath.utils";
+import path from "path";
 
 export type CompareImagesCfg = {
   scaleFactor: number;
@@ -24,21 +30,23 @@ const unlinkSyncSafe = (path: string) =>
 const moveSyncSafe = (pathFrom: string, pathTo: string) =>
   fs.existsSync(pathFrom) && moveFile.sync(pathFrom, pathTo);
 
-export const getScreenshotPathTask = ({
-  title,
-  imagesDir,
-  specPath,
-}: {
-  title: string;
+export const getScreenshotPathInfoTask = (cfg: {
+  titleFromOptions: string;
   imagesDir: string;
   specPath: string;
-}) =>
-  path.join(
-    IMAGE_SNAPSHOT_PREFIX,
-    path.dirname(specPath),
-    ...imagesDir.split("/"),
-    `${sanitize(title)}${FILE_SUFFIX.actual}.png`
-  );
+}) => {
+  const screenshotPath = generateScreenshotPath(cfg);
+
+  return { screenshotPath, title: path.basename(screenshotPath, ".png") };
+};
+
+export const cleanupImagesTask = (config: Cypress.PluginConfigOptions) => {
+  if (config.env["pluginVisualRegressionCleanupUnusedImages"]) {
+    cleanupUnused(config.projectRoot);
+  }
+
+  return null;
+};
 
 export const approveImageTask = ({ img }: { img: string }) => {
   const oldImg = img.replace(FILE_SUFFIX.actual, "");
@@ -65,11 +73,13 @@ export const compareImagesTask = async (
   let error = false;
 
   if (fs.existsSync(cfg.imgOld) && !cfg.updateImages) {
-    const rawImgNew = await importAndScaleImage({
+    const rawImgNewBuffer = await scaleImageAndWrite({
       scaleFactor: cfg.scaleFactor,
       path: cfg.imgNew,
     });
-    const rawImgOld = PNG.sync.read(fs.readFileSync(cfg.imgOld));
+    const rawImgNew = PNG.sync.read(rawImgNewBuffer);
+    const rawImgOldBuffer = fs.readFileSync(cfg.imgOld);
+    const rawImgOld = PNG.sync.read(rawImgOldBuffer);
     const isImgSizeDifferent =
       rawImgNew.height !== rawImgOld.height ||
       rawImgNew.width !== rawImgOld.width;
@@ -108,10 +118,7 @@ export const compareImagesTask = async (
     }
 
     if (error) {
-      fs.writeFileSync(
-        cfg.imgNew.replace(FILE_SUFFIX.actual, FILE_SUFFIX.diff),
-        PNG.sync.write(diff)
-      );
+      writePNG(cfg.imgNew.replace(FILE_SUFFIX.actual, FILE_SUFFIX.diff), diff);
       return {
         error,
         message: messages.join("\n"),
@@ -119,12 +126,19 @@ export const compareImagesTask = async (
         maxDiffThreshold: cfg.maxDiffThreshold,
       };
     } else {
-      // don't overwrite file if it's the same (imgDiff < cfg.maxDiffThreshold && !isImgSizeDifferent)
-      fs.unlinkSync(cfg.imgNew);
+      if (rawImgOld && !isImageCurrentVersion(rawImgOldBuffer)) {
+        writePNG(cfg.imgNew, rawImgNewBuffer);
+        moveFile.sync(cfg.imgNew, cfg.imgOld);
+      } else {
+        // don't overwrite file if it's the same (imgDiff < cfg.maxDiffThreshold && !isImgSizeDifferent)
+        fs.unlinkSync(cfg.imgNew);
+      }
     }
   } else {
     // there is no "old screenshot" or screenshots should be immediately updated
     imgDiff = 0;
+    const imgBuffer = fs.readFileSync(cfg.imgNew);
+    writePNG(cfg.imgNew, imgBuffer);
     moveFile.sync(cfg.imgNew, cfg.imgOld);
   }
 
@@ -151,8 +165,9 @@ export const doesFileExistTask = ({ path }: { path: string }) =>
   fs.existsSync(path);
 
 /* c8 ignore start */
-export const initTaskHook = () => ({
-  [TASK.getScreenshotPath]: getScreenshotPathTask,
+export const initTaskHook = (config: Cypress.PluginConfigOptions) => ({
+  [TASK.getScreenshotPathInfo]: getScreenshotPathInfoTask,
+  [TASK.cleanupImages]: cleanupImagesTask.bind(undefined, config),
   [TASK.doesFileExist]: doesFileExistTask,
   [TASK.approveImage]: approveImageTask,
   [TASK.compareImages]: compareImagesTask,
